+++ src/test/java/org/apache/tomcat/jakartaee/TesterConstants.java
@@ -1,0 +1,9 @@
+package org.apache.tomcat.jakartaee;
+
+public class TesterConstants {
+
+    public static final String JAVA_PRESENT_DOT = "javax.servlet.CommonGatewayInterface";
+    public static final String JAVA_PRESENT_PATH = "javax/servlet/CommonGatewayInterface";
+    public static final String JAVA_NOT_PRESENT_DOT = "javax.servlet.DoesNotExist";
+    public static final String JAVA_NOT_PRESENT_PATH = "javax/servlet/DoesNotExist";
+}