+++ CHANGES.md
@@ -2,7 +2,7 @@
 
 ## 1.0.1 (in progress)
 
-- Fix [#20] by using commons-compression instead of the Java zip code (remm)
+- Fix [#20](https://github.com/apache/tomcat-jakartaee-migration/issues/20) by using commons-compression instead of the Java zip code (remm)
 
 ## 1.0.0